% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_1factor.r
\docType{data}
\name{data_1factor}
\alias{data_1factor}
\title{A sample one factor qPCR dataset}
\format{
A data frame with 9 observations and 6 variables:
\describe{
  \item{SA}{First experimental factor}
  \item{Rep}{Biological replicates}
  \item{EPO}{Mean amplification efficiency of PO gene}
  \item{POCt}{Ct values of PO gene. Each is the mean of technical replicates}
  \item{EGAPDH}{Mean amplification efficiency of GAPDH gene}
  \item{GAPDHCt}{Ct values of GAPDH gene. Each is the mean of technical replicates}
}
}
\source{
Where the data comes from (if applicable)
}
\usage{
data(data_1factor)
}
\description{
A sample dataset for demonstration purposes.
}
\examples{
data_1factor

}
\keyword{datasets}
