% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.r
\docType{data}
\name{data_1factor_one_ref}
\alias{data_1factor_one_ref}
\title{Sample qPCR data (two different conditions)}
\format{
A data frame with 6 observations and 10 variables:
\describe{
  \item{Condition}{Experimental conditions}
  \item{Rep}{Biological replicates}
  \item{C2H2_26_E}{amplification efficiency of C2H2_26 gene}
  \item{C2H2_26_Ct}{Ct values of C2H2_26 gene. Each is the mean of technical replicates}
  \item{C2H2_01_E}{Amplification efficiency of C2H2_01 gene}
  \item{C2H2_01_Ct}{Ct values of C2H2_01 gene. Each is the mean of technical replicates}
  \item{C2H2_12_E}{Amplification efficiency of C2H2_12 gene}
  \item{C2H2_12_Ct}{Ct values of C2H2_12 gene}
  \item{ref_E}{Amplification efficiency of ref gene}
  \item{ref_Ct}{Ct values of ref gene}
}
}
\source{
University of Kurdistan
}
\usage{
data_1factor_one_ref
}
\description{
Sample qPCR data (two different conditions)
}
\keyword{internal}
