% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.r
\docType{data}
\name{data_1factor_one_ref_Eff}
\alias{data_1factor_one_ref_Eff}
\title{Sample qPCR data (two different conditions)}
\format{
A data frame with 6 observations and 6 variables:
\describe{
  \item{Con}{Experimental conditions}
  \item{r}{Biological replicates}
  \item{target_E}{Amplification efficiency of target gene}
  \item{target_Ct}{Ct values of target gene}
  \item{Actin_E}{Amplification efficiency of reference gene}
  \item{Actin_Ct}{Ct values of reference gene}
}
}
\source{
University of Kurdistan
}
\usage{
data_1factor_one_ref_Eff
}
\description{
Sample qPCR data (two different conditions)
}
\keyword{internal}
