% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.r
\docType{data}
\name{data_1factor_Two_ref}
\alias{data_1factor_Two_ref}
\title{Sample qPCR data (one target and two reference genes under two different conditions)}
\format{
A data frame with 6 observations and 8 variables:
\describe{
  \item{Condition}{Experimental conditions}
  \item{Rep}{Biological replicates}
  \item{E_DER5}{amplification efficiency of DER5 gene}
  \item{Ct_DER5}{Ct values of DER5 gene. Each is the mean of technical replicates}
  \item{E_Actin}{Amplification efficiency of Actin gene}
  \item{Ct_Actin}{Ct values of Actin gene. Each is the mean of technical replicates}
  \item{E_HPRT}{Amplification efficiency of HPRT gene}
  \item{Ct_HPRT}{Ct values of HPRT gene}
}
}
\source{
Not applicable
}
\usage{
data_1factor_Two_ref
}
\description{
Sample qPCR data (one target and two reference genes under two different conditions)
}
\keyword{internal}
