% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{data_2factor}
\alias{data_2factor}
\title{Sample data (two factor)}
\format{
A data frame with 18 observations and 7 variables:
\describe{
  \item{Genotype}{First experimental factor}
  \item{Drought}{Second experimental factor}
  \item{Rep}{Biological replicates}
  \item{PO}{Mean amplification efficiency of PO gene}
  \item{Ct_PO}{Ct values of PO gene. Each is the mean of technical replicates}
  \item{GAPDH}{Mean amplification efficiency of GAPDH gene}
  \item{Ct_GAPDH}{Ct values of GAPDH gene. Each is the mean of technical replicates}
}
}
\source{
Not applicable
}
\usage{
data_2factor
}
\description{
A sample dataset. Each line belongs to a separate individual (non-repeated measure experiment).
}
\keyword{internal}
