% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.r
\docType{data}
\name{data_2factor3ref}
\alias{data_2factor3ref}
\title{Sample data in (two factor with blocking factor)}
\format{
A data frame with 18 observations and 8 variables:
\describe{
  \item{factor1}{First experimental factor}
  \item{factor2}{Second experimental factor}
  \item{Rep}{Biological replicates}
  \item{E_PO}{Mean amplification efficiency of PO gene}
  \item{Ct_PO}{Ct values of PO gene. Each is the mean of technical replicates}
  \item{E_GAPDH}{Mean amplification efficiency of GAPDH gene}
  \item{Ct_GAPDH}{Ct values of GAPDH gene. Each is the mean of technical replicates}
  \item{ref2E}{Mean amplification efficiency of ref2 gene}
  \item{ref2Ct}{Ct values of ref2 gene. Each is the mean of technical replicates}
  \item{ref3E}{Mean amplification efficiency of ref3 gene}
  \item{ref3Ct}{Ct values of GAPDH gene. Each is the mean of technical replicates}
}
}
\source{
Not applicable
}
\usage{
data_2factor3ref
}
\description{
A sample qPCR data set with blocking factor. Each line belongs to a separate individual (non-repeated measure experiment).
}
\keyword{internal}
