% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{data_2factorBlock}
\alias{data_2factorBlock}
\title{Sample data in (two factor with blocking factor)}
\format{
A data frame with 18 observations and 8 variables:
\describe{
  \item{Type}{First experimental factor}
  \item{Concentration}{Second experimental factor}
  \item{block}{Second experimental factor}
  \item{Rep}{Biological replicates}
  \item{PO}{Mean amplification efficiency of PO gene}
  \item{Ct_PO}{Ct values of PO gene. Each is the mean of technical replicates}
  \item{GAPDH}{Mean amplification efficiency of GAPDH gene}
  \item{Ct_GAPDH}{Ct values of GAPDH gene. Each is the mean of technical replicates}
}
}
\source{
Not applicable
}
\usage{
data_2factorBlock
}
\description{
A sample qPCR data set with blocking factor. Each line belongs to a separate individual (non-repeated measure experiment).
}
\keyword{internal}
