% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{data_2factorBlock3ref}
\alias{data_2factorBlock3ref}
\title{Sample data in (two factor with blocking factor and 3 reference genes)}
\format{
A data frame with 18 observations and 8 variables:
\describe{
  \item{Type}{First experimental factor}
  \item{Concentration}{Second experimental factor}
  \item{block}{blocking factor}
  \item{Rep}{Biological replicates}
  \item{PO}{Mean amplification efficiency of PO gene}
  \item{Ct_PO}{Ct values of PO gene. Each is the mean of technical replicates}
  \item{NLM}{Mean amplification efficiency of NLM gene}
  \item{Ct_NLM}{Ct values of NLM gene. Each is the mean of technical replicates}
  \item{ref1}{Mean amplification efficiency of ref1 gene}
  \item{Ct_ref1}{Ct values of ref1 gene. Each is the mean of technical replicates}
  \item{ref2}{Mean amplification efficiency of ref2 gene}
  \item{Ct_ref2}{Ct values of ref2 gene. Each is the mean of technical replicates}
  \item{ref3}{Mean amplification efficiency of ref3 gene}
  \item{Ct_ref3}{Ct values of GAPDH gene. Each is the mean of technical replicates}
}
}
\source{
Not applicable
}
\usage{
data_2factorBlock3ref
}
\description{
A sample qPCR data set with blocking factor and 3 reference genes. Each line belongs to a separate individual (non-repeated measure experiment).
}
\keyword{external}
