% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{data_efficiency1}
\alias{data_efficiency1}
\title{Sample qPCR data: amplification efficiency 1}
\format{
A data frame with 21 observations and 4 variables:
\describe{
  \item{Dilutions}{Dilution factor}
  \item{C2H2.26}{Target gene 1}
  \item{C2H2.01}{Target gene 2}
  \item{GAPDH}{Reference gene}
}
}
\source{
Not applicable
}
\usage{
data_efficiency1
}
\description{
A sample qPCR dataset for demonstrating efficiency calculation.
}
\keyword{internal}
