% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{data_efficiency_Yuan2006PMCBioinf}
\alias{data_efficiency_Yuan2006PMCBioinf}
\title{Sample qPCR data: amplification efficiency 2}
\format{
A data frame with 12 observations and 5 variables:
\describe{
  \item{dilutions}{dilution factor}
  \item{ref_control}{reference gene in control condition}
  \item{ref_treatment}{reference gene in treatment condition}
  \item{target_control}{target gene in control condition}
  \item{target_treatment}{Reference gene in treatment condition}
}
}
\source{
Yuan2006PMCBioinf
}
\usage{
data_efficiency_Yuan2006PMCBioinf
}
\description{
A sample qPCR dataset for demonstrating efficiency calculation.
}
\keyword{internal}
