% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{data_Heffer2020PlosOne}
\alias{data_Heffer2020PlosOne}
\title{Sample data in (multi-group design with 6 target genes)}
\format{
A data frame with 18 observations and 16 variables:
\describe{
  \item{Treatment}{Experimental treatment group}
  \item{rep}{Biological replicates}

  \item{Fn1, Ct.Fn1}{Efficiency and Ct values of Fn1 gene}
  \item{Col1a1, Ct.Col1a1}{Efficiency and Ct values of Col1a1 gene}
  \item{Acta2, Ct.Acta2}{Efficiency and Ct values of Acta2 gene}
  \item{TgfB, Ct.TgfB}{Efficiency and Ct values of TgfB gene}
  \item{Tnfa, Ct.Tnfa}{Efficiency and Ct values of Tnfa gene}
  \item{Mcp1, Ct.Mcp1}{Efficiency and Ct values of Mcp1 gene}
  \item{Gapdh, Ct.Gapdh}{Efficiency and Ct values of Gapdh reference gene}
}
}
\source{
Not applicable
}
\usage{
data_Heffer2020PlosOne
}
\description{
A sample qPCR data set with four experimental groups
(Uninjected, SNConly, DMSORPE, SNCRPE) and six target genes
normalized to one reference gene (Gapdh).
Each line belongs to a separate individual (non-repeated measure experiment).
}
\details{
This dataset is suitable for t-test or one-way ANOVA based
expression analysis of multiple genes.
}
\keyword{internal}
