% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.r
\docType{data}
\name{data_repeated_measure_2}
\alias{data_repeated_measure_2}
\title{Repeated measure sample data}
\format{
A data frame with 18 observations and 7 variables:
\describe{
  \item{treatment}{treatment}
  \item{time}{time course levels}
  \item{id}{experimental factor}
  \item{Target}{Amplification efficiency of target gene}
  \item{Target_Ct}{Ct of target gene}
  \item{Ref}{Amplification efficiency of reference gene}
  \item{Ref_Ct}{Ct of reference gene}
}
}
\source{
NA
}
\usage{
data_repeated_measure_2
}
\description{
A repeated measure sample data in which 6 individuals have been analysed.  In the "id" column, a unique number is assigned to each individual,  e.g. all the three number 1 indicate one individual.
samples are taken or measurements are scored over different time points (time column) from each individual.
}
\keyword{internal}
