% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{data_repeated_measure_3bLock}
\alias{data_repeated_measure_3bLock}
\title{Sample data in (repeated-measures design with two reference genes)}
\format{
A data frame with 18 observations and 10 variables:
\describe{
  \item{treatment}{Experimental treatment (untreated, treated)}
  \item{time}{Time point of measurement}
  \item{blk}{Blocking factor}
  \item{id}{Subject/individual identifier for repeated measures}

  \item{Target}{Mean amplification efficiency of target gene}
  \item{Ct_Target}{Ct values of target gene}

  \item{Ref1}{Mean amplification efficiency of reference gene 1}
  \item{Ct_Ref1}{Ct values of reference gene 1}

  \item{Ref2}{Mean amplification efficiency of reference gene 2}
  \item{Ct_Ref2}{Ct values of reference gene 2}
}
}
\source{
Not applicable
}
\usage{
data_repeated_measure_3bLock
}
\description{
A sample qPCR data set with repeated measurements over time,
a blocking factor, and two reference genes. The experiment includes
two treatments (untreated and treated) with repeated measures
on the same individuals (id).
}
\details{
This dataset is suitable for repeated-measures mixed-model analysis
and normalization using multiple reference genes.
}
\keyword{internal}
