% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.r
\docType{data}
\name{data_ttest}
\alias{data_ttest}
\title{Sample qPCR data of an experiment conducted under two different conditions}
\format{
A data frame with 24 observations and 4 variables:
\describe{
  \item{Condition}{Experimental conditions}
  \item{Gene}{Genes}
  \item{E}{Amplification efficiency}
  \item{Ct}{Ct values}
}
}
\source{
University of Kurdistan
}
\usage{
data_ttest
}
\description{
A sample data for demonstrating qPCR data analysis. In unpaired condition, each line belongs to a separate individual.
However t-test can be applied for paired samples where the data is acquired from same individuals in two conditions e.g. before and after a treatment.
}
\keyword{internal}
