% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{data_ttest18genes}
\alias{data_ttest18genes}
\title{Sample data in (two-group design with multiple target genes)}
\format{
A data frame with 8 observations and 38 variables:
\describe{
  \item{Condition}{Experimental group (Control, Treatment)}
  \item{Rep}{Biological replicates}

  \item{ANGPT1, Ct.ANGPT1}{Efficiency and Ct values of ANGPT1 gene}
  \item{ANGPT2, Ct.ANGPT2}{Efficiency and Ct values of ANGPT2 gene}
  \item{CCL2, Ct.CCL2}{Efficiency and Ct values of CCL2 gene}
  \item{CCL5, Ct.CCL5}{Efficiency and Ct values of CCL5 gene}
  \item{CSF2, Ct.CSF2}{Efficiency and Ct values of CSF2 gene}
  \item{FGF2, Ct.FGF2}{Efficiency and Ct values of FGF2 gene}
  \item{IL1A, Ct.IL1A}{Efficiency and Ct values of IL1A gene}
  \item{IL1B, Ct.IL1B}{Efficiency and Ct values of IL1B gene}
  \item{IL6, Ct.IL6}{Efficiency and Ct values of IL6 gene}
  \item{IL8, Ct.IL8}{Efficiency and Ct values of IL8 gene}
  \item{PDGFA, Ct.PDGFA}{Efficiency and Ct values of PDGFA gene}
  \item{PDGFB, Ct.PDGFB}{Efficiency and Ct values of PDGFB gene}
  \item{TGFA, Ct.TGFA}{Efficiency and Ct values of TGFA gene}
  \item{TGFB, Ct.TGFB}{Efficiency and Ct values of TGFB gene}
  \item{TNF, Ct.TNF}{Efficiency and Ct values of TNF gene}
  \item{VEGFA, Ct.VEGFA}{Efficiency and Ct values of VEGFA gene}
  \item{VEGFB, Ct.VEGFB}{Efficiency and Ct values of VEGFB gene}
  \item{VEGFC, Ct.VEGFC}{Efficiency and Ct values of VEGFC gene}
  \item{GAPDH, Ct.GAPDH}{Efficiency and Ct values of GAPDH reference gene}
}
}
\source{
Not applicable
}
\usage{
data_ttest18genes
}
\description{
A sample qPCR data set with two experimental groups (Control and Treatment)
and 18 target genes normalized to one reference gene (GAPDH).
Each line belongs to a separate individual (non-repeated measure experiment).
This dataset is suitable for two-sample t-test based expression analysis.
}
\keyword{internal}
