% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ttest2.r
\docType{data}
\name{data_ttest2}
\alias{data_ttest2}
\title{Sample data (one target, two reference)}
\format{
A data frame with 18 observations and 4 variables:
\describe{
  \item{Condition}{Experimental conditions}
  \item{Gene}{Genes}
  \item{E}{Amplification efficiency}
  \item{Ct}{Ct values}
}
}
\source{
Na
}
\usage{
data(data_ttest2)
}
\description{
One target and two reference gens for demonstrating qPCR data analysis.
}
\examples{
data(data_ttest2)
data_ttest2

}
\keyword{datasets}
