% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{data_withTechRep}
\alias{data_withTechRep}
\title{Sample data (with technical replicates)}
\format{
A data frame with 18 observations and 9 variables:
\describe{
  \item{Condition}{experimental factor}
  \item{biolrep}{biological replicate}
  \item{techrep}{technical replicates}
  \item{target}{Amplification efficiency of target gene}
  \item{Ct_target}{Ct of target gene}
  \item{ref}{Amplification efficiency of reference gene}
  \item{Ct_ref}{Ct of reference gene}
}
}
\source{
Not applicable
}
\usage{
data_withTechRep
}
\description{
A sample data for calculating biological replicated. Each line belongs to a separate individual (non-repeated measure experiment).
}
\keyword{internal}
