% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{data_Yuan2006PMCBioinf}
\alias{data_Yuan2006PMCBioinf}
\title{Sample data in (one factor with one reference gene)}
\format{
A data frame with 24 observations and 6 variables:
\describe{
  \item{condition}{Experimental factor with two levels (control, treatment)}
  \item{rep}{Biological replicates}
  \item{target}{Mean amplification efficiency of target gene}
  \item{Ct_target}{Ct values of target gene. Each is the mean of technical replicates}
  \item{ref}{Mean amplification efficiency of reference gene}
  \item{Ct_ref}{Ct values of reference gene. Each is the mean of technical replicates}
}
}
\source{
Yuan2006PMCBioinf
}
\usage{
data_Yuan2006PMCBioinf
}
\description{
A sample qPCR data set with one experimental factor (condition)
and one reference gene. Each line belongs to a separate individual
(non-repeated measure experiment).
}
\keyword{internal}
