% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcrhlpr.r
\name{.theme_pub}
\alias{.theme_pub}
\title{Publication-ready ggplot2 theme}
\usage{
.theme_pub(
  base_size = 12,
  base_family = "sans",
  legend_position = "right",
  ...
)
}
\arguments{
\item{base_size}{Numeric. Base font size (default 12)}

\item{base_family}{Character. Font family (default "sans")}

\item{legend_position}{Character or numeric vector. Position of legend (default "right")}

\item{...}{Additional theme arguments}
}
\value{
ggplot2 theme object
}
\description{
Publication-ready ggplot2 theme
}
