% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalVariables.r
\name{globalVariables}
\alias{globalVariables}
\title{Internal global variables}
\description{
These objects are declared to avoid R CMD check notes about
non-standard evaluation (e.g., in ggplot2).
}
\keyword{internal}
