% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcrANOVA.r
\name{qpcrANOVA}
\alias{qpcrANOVA}
\title{qPCR Analysis of Variance}
\usage{
qpcrANOVA(
  x,
  numberOfrefGenes = 1,
  block = NULL,
  p.adj = c("none", "holm", "hommel", "hochberg", "bonferroni", "BH", "BY", "fdr")
)
}
\arguments{
\item{x}{A data frame consisting of condition columns, target gene efficiency (E), target Gene Ct, reference gene efficiency and reference gene Ct values, respectively. Each Ct in the data frame is the mean of technical replicates. Complete amplification efficiencies of 2 was assumed in the example data for all wells but the calculated efficienies can be used instead.}

\item{numberOfrefGenes}{number of reference genes. Up to two reference genes can be handled.}

\item{block}{column name of the blocking factor (for correct column arrangement see example data.)}

\item{p.adj}{Method for adjusting p values (see p.adjust)}
}
\value{
A list with 5 elements:
\describe{
  \item{Final_data}{The final row data including weighed delta Ct (wDCt) values.}
  \item{lm}{The output of linear model analysis including ANOVA tables based on factorial experiment and completely randomized design (CRD).}
  \item{ANOVA_factorial}{ANOVA table based on factorial arrangement}
  \item{ANOVA_CRD}{ANOVA table based on CRD}
  \item{Result}{The main result table including treatments and factors, wDCt, LCL, UCL, letters and standard deviation of average relative expression data.}
  \item{Post_hoc_Test}{Post hoc testing of means along with the significance and confidence interval.}
}
}
\description{
Analysis of Variance of qPCR data
}
\details{
The qpcrANOVA performs ANOVA (analysis of variance) of qPCR data. It is suitable when there is a factor with more than two levels or when the are more that a condition factor in the experiment or when there is a blocking factor.
}
\examples{

# If the data include technical replicates, means of technical replicates
# should be calculated first using meanTech function.

# Applying ANOVA analysis
qpcrANOVA(
     data_3factor_a,
     numberOfrefGenes = 1,
     p.adj = "none")


qpcrANOVA(
    data_2factorBlock,
    block = "Block",
    numberOfrefGenes = 1)


}
\references{
Livak, Kenneth J, and Thomas D Schmittgen. 2001. Analysis of
Relative Gene Expression Data Using Real-Time Quantitative PCR and the
Double Delta CT Method. Methods 25 (4). doi:10.1006/meth.2001.1262.

Ganger, MT, Dietz GD, and Ewing SJ. 2017. A common base method for analysis of qPCR data
and the application of simple blocking in qPCR experiments. BMC bioinformatics 18, 1-11.

Yuan, Joshua S, Ann Reed, Feng Chen, and Neal Stewart. 2006.
Statistical Analysis of Real-Time PCR Data. BMC Bioinformatics 7 (85). doi:10.1186/1471-2105-7-85.
}
\author{
Ghader Mirzaghaderi
}
