% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REPEATED_DDCt.R
\name{REPEATED_DDCt}
\alias{REPEATED_DDCt}
\title{Delta Delta Ct repeated measure analysis}
\usage{
REPEATED_DDCt(
  x,
  numOfFactors,
  numberOfrefGenes,
  mainFactor.column,
  block,
  mainFactor.level.order = NULL,
  p.adj = "none",
  analyseAllTarget = TRUE
)
}
\arguments{
\item{x}{The input data frame containing experimental design columns, target gene
E/Ct column pairs, and reference gene E/Ct column pairs. Reference gene
columns must be located at the end of the data frame.}

\item{numOfFactors}{Integer. Number of experimental factor columns
(excluding \code{rep} and optional \code{block}).}

\item{numberOfrefGenes}{Integer. Number of reference genes.}

\item{mainFactor.column}{Integer. Column index of the factor (commonly \code{"time"}) for which relative expression is calculated.}

\item{block}{Character or \code{NULL}. Name of the blocking factor column.
When a qPCR experiment is done in multiple qPCR plates, 
variation resulting from the plates may interfere with the actual amount of 
gene expression. One solution is to conduct each plate as a randomized block 
so that at least one replicate of each treatment and control is present 
on a plate. Block effect is usually considered as random and its interaction 
with any main effect is not considered.}

\item{mainFactor.level.order}{Optional character vector specifying the order of levels for the main factor.
If \code{NULL}, the first observed level is used as the calibrator.
If provided, the first element of the vector is used as the calibrator level.}

\item{p.adj}{Method for p-value adjustment. See \code{\link[stats]{p.adjust}}.}

\item{analyseAllTarget}{Logical or character.
If \code{TRUE} (default), all target genes are analysed.
Alternatively, a character vector specifying the names (names of their Efficiency columns) of target genes
to be analysed.}
}
\value{
An object containing expression table, lm model, residuals, raw data and ANOVA table for each gene:
\describe{  
\item{ddCt expression table along with per-gene statistical comparison outputs}{\code{object$relativeExpression}}
\item{ANOVA table}{\code{object$perGene$gene_name$ANOVA_table}}
\item{lm ANOVA}{\code{object$perGene$gene_name$lm}}
\item{lm_formula}{\code{object$perGene$gene_name$lm_formula}}
\item{Residuals}{\code{resid(object$perGene$gene_name$lm)}}
}
}
\description{
\code{REPEATED_DDCt} function performs Delta Delta Ct (ddCt) method
analysis of observations repeatedly taken over different time courses. 
Data may be obtained over time from a uni- or multi-factorial experiment.
}
\details{
ddCt analysis of repeated measure data is performed for 
the \code{mainFactor.column} based on a full model factorial experiment.

All the functions for relative expression analysis (including `TTEST_DDCt()`, 
`WILCOX_DDCt()`, `ANOVA_DDCt()`, `ANCOVA_DDCt()`, `REPEATED_DDCt()`, and `ANOVA_DCt()`) return the 
relative expression table which include fold change and corresponding 
statistics. The output of `ANOVA_DDCt()`, `ANCOVA_DDCt()`, `ANCOVA_DDCt()`, `REPEATED_DDCt()`, 
and `ANOVA_DCt()` also include lm models, residuals, raw data and ANOVA table 
for each gene. 

The expression table returned by `TTEST_DDCt()`, 
`WILCOX_DDCt()`, `ANOVA_DDCt()`, `ANCOVA_DDCt()`, and `REPEATED_DDCt()` functions 
include these columns: gene (name of target genes), 
contrast (calibrator level and contrasts for which the relative expression is 
computed), ddCt (mean of weighted delta delta Ct values), RE (relative expression or 
fold change = 2^-ddCt),  log2FC (log(2) of relative expression or fold change), 
pvalue, sig (per-gene significance), LCL (95\% lower confidence level), UCL (95\% upper confidence level),
se (standard error of mean calculated from the weighted delta Ct values of each of the main factor levels),
Lower.se.RE (The lower limit error bar for RE which is 2^(log2(RE) - se)), 
Upper.se.RE (The upper limit error bar for RE which is 2^(log2(RE) + se)),
Lower.se.log2FC (The lower limit error bar for log2 RE), and 
Upper.se.log2FC (The upper limit error bar for log2 RE)
}
\examples{
data1 <- read.csv(system.file("extdata", "data_repeated_measure_1.csv", package = "rtpcr"))
REPEATED_DDCt(
  data1,
  numOfFactors = 1,
  numberOfrefGenes = 1,
  mainFactor.column = 1,
  block = NULL)



data2 <- read.csv(system.file("extdata", "data_repeated_measure_2.csv", package = "rtpcr"))
REPEATED_DDCt(
  data2,
  numOfFactors = 2,
  numberOfrefGenes = 1,
  mainFactor.column = 2, 
  block = NULL,
  p.adj = "none")
          
}
\references{
LivakKJ, Schmittgen TD (2001).
Analysis of Relative Gene Expression Data Using Real-Time Quantitative PCR
and the Double Delta CT Method.
\emph{Methods}, 25(4), 402–408.
doi:10.1006/meth.2001.1262

Ganger MT, Dietz GD, and Ewing SJ (2017).
A common base method for analysis of qPCR data and the application of
simple blocking in qPCR experiments.
\emph{BMC Bioinformatics}, 18, 1–11.

Taylor SC, Nadeau K, Abbasi M, Lachance C, Nguyen M, Fenrich, J. (2019). 
The ultimate qPCR experiment: producing publication quality, reproducible 
data the first time. \emph{Trends in Biotechnology}, 37, 761-774. 

Yuan JS, Reed A, Chen F, Stewart N (2006).
Statistical Analysis of Real-Time PCR Data.
\emph{BMC Bioinformatics}, 7, 85.
}
