% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the Shiny app}
\usage{
run_app(display.mode = "normal")
}
\arguments{
\item{display.mode}{Character. Passed to shiny::runApp().
Default is "normal".}
}
\value{
Invisibly returns the Shiny app object.
}
\description{
Starts the interactive Shiny web application included in this package.
}
