% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threeFACTORplot.r
\name{threeFACTORplot}
\alias{threeFACTORplot}
\title{Bar plot of the relative gene expression (RE) from a three-factor experiment}
\usage{
threeFACTORplot(
  x,
  arrangement = c(1, 2, 3),
  bar.width = 0.5,
  fill = "Reds",
  xlab = "Gene",
  ylab = "Relative Expression",
  errorbar = "std",
  y.axis.adjust = 0.5,
  y.axis.by = 2,
  letter.position.adjust = 0.3,
  legend.title = "Legend Title",
  legend.position = c(0.4, 0.8),
  fontsize = 12,
  fontsizePvalue = 7,
  show.letters = TRUE
)
}
\arguments{
\item{x}{a data frame created by \link{qpcrANOVA} function via running \code{res <- qpcrANOVA(x)$Result} on a three factorial data such as \code{datathreegr} example data.}

\item{arrangement}{arrangement of the grouping columns in the output graph, for example c(2,3,1) or c(1,3,2). This affects grapg output.}

\item{bar.width}{a positive number determining bar width.}

\item{fill}{a color vector specifying the fill color for the columns of the bar plot. One of the palettes in \code{\link[RColorBrewer]{display.brewer.all}} (e.g. "Reds" or "Blues", ...) can be applied.}

\item{xlab}{the title of the x axis}

\item{ylab}{the title of the y axis}

\item{errorbar}{Type of error bar, can be \code{std} or \code{ci}.}

\item{y.axis.adjust}{a negative or positive number for reducing or increasing the length of the y axis.}

\item{y.axis.by}{determines y axis step length}

\item{letter.position.adjust}{adjust the distance between the grouping letters to the error bars}

\item{legend.title}{legend title}

\item{legend.position}{a two digit vector specifying the legend position.}

\item{fontsize}{all fonts size of the plot}

\item{fontsizePvalue}{font size of the pvalue labels}

\item{show.letters}{a logical variable. If TRUE, mean grouping letters are added to the bars.}
}
\value{
Bar plot of the average fold change for target genes along with the significance and the 95\% confidence interval as error bars.
}
\description{
Bar plot of the relative expression (RE) of a gene along with the confidence interval and significance
}
\details{
The \code{threeFACTORplot} function generates the bar plot of the average fold change for target genes along with the significance and the 95\% confidence interval as error bars.
}
\examples{

#' # See a sample data frame
data_3factor_b

# Before plotting, the result needs to be extracted as below:
res <- qpcrANOVA(data_3factor_b, numberOfrefGenes = 1)$Result
res

# Arrange the first three colunms of the result table.
# This determines the columns order and shapes the plot output.
threeFACTORplot(res,
    arrangement = c(3, 1, 2),
    xlab = "condition")



threeFACTORplot(res,
   arrangement = c(1, 2, 3),
   bar.width = 0.5,
   fill = "Greys",
   xlab = "Genotype",
   ylab = "Relative Expression")



# Reordering factor levels to a desired order.
res$Conc <- factor(res$Conc, levels = c("L","M","H"))
res$Type <- factor(res$Type, levels = c("S","R"))

# Producing the plot
threeFACTORplot(res,
   arrangement = c(2, 3, 1),
   bar.width = 0.5,
   fill = "Reds",
   xlab = "Drought",
   ylab = "Relative Expression",
   errorbar = "std",
   legend.title = "Genotype",
   legend.position = c(0.2, 0.8))


# When using ci as error, increase the 
# y.axis.adjust value to see the plot correctly!
threeFACTORplot(res,
   arrangement = c(2, 3, 1),
   bar.width = 0.8,
   fill = "Greens",
   xlab = "Drought",
   ylab = "Relative Expression",
   errorbar = "ci",
   y.axis.adjust = 8,
   y.axis.by = 2,
   letter.position.adjust = 0.6,
   legend.title = "Genotype",
   fontsize = 12,
   legend.position = c(0.2, 0.8),
   show.letters = TRUE)



}
\author{
Ghader Mirzaghaderi
}
