% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoFACTORplot.r
\name{twoFACTORplot}
\alias{twoFACTORplot}
\title{Bar plot of the relative gene expression from a two-factor experiment}
\usage{
twoFACTORplot(
  x,
  x.axis.factor,
  group.factor,
  width = 0.5,
  fill = "Blues",
  y.axis.adjust = 0.5,
  y.axis.by = 2,
  errorbar = "std",
  show.letters = TRUE,
  letter.position.adjust = 0.1,
  ylab = "Relative Expression",
  xlab = "Gene",
  legend.position = c(0.09, 0.8),
  fontsize = 12
)
}
\arguments{
\item{x}{a data frame created by \link{qpcrANOVA} function via running \code{res <- qpcrANOVA(x)$Result}.}

\item{x.axis.factor}{x-axis factor.}

\item{group.factor}{grouping factor.}

\item{width}{a positive number determining bar width.}

\item{fill}{specify the fill color vector for the columns of the bar plot. One of the palettes in \code{\link[RColorBrewer]{display.brewer.all}} (e.g. "Reds" or "Blues", ...) can be applied.}

\item{y.axis.adjust}{a negative or positive number for reducing or increasing the length of the y axis.}

\item{y.axis.by}{determines y axis step length.}

\item{errorbar}{Type of error bar, can be \code{std} or \code{ci}.}

\item{show.letters}{a logical variable. If TRUE, mean grouping letters are added to the bars.}

\item{letter.position.adjust}{adjust the distance between the grouping letters to the error bars.}

\item{ylab}{the title of the y axis.}

\item{xlab}{the title of the x axis.}

\item{legend.position}{a two digit vector specifying the legend position.}

\item{fontsize}{size of all fonts  of the plot.}
}
\value{
Bar plot of the average fold change for target genes along with the significance and the 95\% confidence interval as error bars.
}
\description{
Bar plot of the relative expression of a gene along with the 95\% confidence interval and significance
}
\details{
The \code{twoFACTORplot} function generates the bar plot of the average fold change for target genes along with the significance and the 95\% confidence interval as error bars.
}
\examples{

# Before plotting, the result needs to be extracted as below:
res <- qpcrANOVA(data_2factor)$Result
res

# Plot of the 'res' data with 'Genotype' as grouping factor
twoFACTORplot(res,
   x.axis.factor = Drought,
   group.factor = Genotype,
   width = 0.5,
   fill = "Greens",
   y.axis.adjust = 1,
   y.axis.by = 2,
   ylab = "Relative Expression",
   xlab = "Drought Levels",
   show.letters = TRUE, 
   letter.position.adjust = 0.3,
   legend.position = c(0.09, 0.8),
   errorbar = "ci")

# Plotting the same data with 'Drought' as grouping factor
twoFACTORplot(res,
   x.axis.factor = Genotype,
   group.factor = Drought,
   xlab = "Genotype",
   fill = "Blues",
   fontsize = 12,
   show.letters = FALSE)


}
\author{
Ghader Mirzaghaderi
}
