% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2factor.r
\docType{data}
\name{data_2factor}
\alias{data_2factor}
\title{A sample two factor qPCR dataset}
\format{
A data frame with 18 observations and 7 variables:
\describe{
  \item{Genotype}{First experimental factor}
  \item{Drought}{Second experimental factor}
  \item{Rep}{Biological replicates}
  \item{EPO}{Mean amplification efficiency of PO gene}
  \item{POCt}{Ct values of PO gene. Each is the mean of technical replicates}
  \item{EGAPDH}{Mean amplification efficiency of GAPDH gene}
  \item{GAPDHCt}{Ct values of GAPDH gene. Each is the mean of technical replicates}
}
}
\source{
Where the data comes from (if applicable)
}
\usage{
data(data_2factor)
}
\description{
A sample dataset for demonstration purposes.
}
\examples{
data(data_2factor)
data_2factor

}
\keyword{datasets}
