% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.r
\docType{data}
\name{data_2factor}
\alias{data_2factor}
\title{Sample data (two factor)}
\format{
A data frame with 18 observations and 7 variables:
\describe{
  \item{Genotype}{First experimental factor}
  \item{Drought}{Second experimental factor}
  \item{Rep}{Biological replicates}
  \item{EPO}{Mean amplification efficiency of PO gene}
  \item{POCt}{Ct values of PO gene. Each is the mean of technical replicates}
  \item{EGAPDH}{Mean amplification efficiency of GAPDH gene}
  \item{GAPDHCt}{Ct values of GAPDH gene. Each is the mean of technical replicates}
}
}
\source{
Not applicable
}
\usage{
data_2factor
}
\description{
A sample dataset for demonstration purposes.
}
\keyword{datasets}
