% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.r
\docType{data}
\name{data_2factorBlock}
\alias{data_2factorBlock}
\title{Sample data (two factor with blocking factor)}
\format{
A data frame with 18 observations and 8 variables:
\describe{
  \item{factor1}{First experimental factor}
  \item{factor2}{Second experimental factor}
  \item{block}{Second experimental factor}
  \item{Rep}{Biological replicates}
  \item{EPO}{Mean amplification efficiency of PO gene}
  \item{POCt}{Ct values of PO gene. Each is the mean of technical replicates}
  \item{EGAPDH}{Mean amplification efficiency of GAPDH gene}
  \item{GAPDHCt}{Ct values of GAPDH gene. Each is the mean of technical replicates}
}
}
\source{
Not applicable
}
\usage{
data_2factorBlock
}
\description{
A sample qPCR data set with blocking factor.
}
\keyword{datasets}
