% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.r
\docType{data}
\name{data_3factor_b}
\alias{data_3factor_b}
\title{Sample data (three factor)}
\format{
A data frame with 36 observations and 8 variables:
\describe{
  \item{Type}{The first experimental factor}
  \item{Conc}{The second experimental factor}
  \item{SA}{The third experimental factor}
  \item{Replicate}{Biological replicates}
  \item{EPO}{Mean amplification efficiency of PO gene}
  \item{POCt}{Ct values of PO gene. Each is the mean of technical replicates}
  \item{EGAPDH}{Mean amplification efficiency of GAPDH gene}
  \item{GAPDHCt}{Ct values of GAPDH gene. Each is the mean of technical replicates}
}
}
\source{
University of Kurdistan
}
\usage{
data_3factor_b
}
\description{
A sample real time PCR data for demonstration purposes.
}
\keyword{datasets}
