% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.r
\docType{data}
\name{data_efficiency}
\alias{data_efficiency}
\title{Sample data (amplification efficiency)}
\format{
A data frame with 21 observations and 3 variables:
\describe{
  \item{dilutions}{Dilution factor}
  \item{C2H2.26}{Target gene}
  \item{GAPDH}{Reference gene}
}
}
\source{
Where the data comes from (if applicable)
}
\usage{
data_efficiency
}
\description{
A sample qPCR dataset for demonstrating efficiency calculation.
}
\keyword{datasets}
