% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.r
\docType{data}
\name{data_ttest}
\alias{data_ttest}
\title{Sample data (one factor-two level qPCR)}
\format{
A data frame with 24 observations and 4 variables:
\describe{
  \item{Condition}{Experimental conditions}
  \item{Gene}{Genes}
  \item{E}{Amplification efficiency}
  \item{Ct}{Ct values}
}
}
\source{
University of Kurdistan
}
\usage{
data_ttest
}
\description{
A sample data for demonstrating qPCR data analysis.
}
\keyword{datasets}
