% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_withTechRep.r
\docType{data}
\name{data_withTechRep}
\alias{data_withTechRep}
\title{Sample data (with technical replicates)}
\format{
A data frame with 18 observations and 9 variables:
\describe{
  \item{factor1}{experimental factor}
  \item{factor2}{experimental factor}
  \item{factor3}{experimental factor}
  \item{biolrep}{biological replicate}
  \item{techrep}{technical replicates}
  \item{Eref}{Amplification efficiency of reference gene}
  \item{refCt}{Ct of refence gene}
  \item{Etarget}{Amplification efficiency of target gene}
  \item{targetCt}{Ct of target gene}
}
}
\source{
Where the data comes from (if applicable)
}
\usage{
data(data_withTechRep)
}
\description{
A sample data for calculating biological replicated.
}
\examples{
data(data_withTechRep)
data_withTechRep

}
\keyword{datasets}
