% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.r
\docType{data}
\name{data_withTechRep}
\alias{data_withTechRep}
\title{Sample data (with technical replicates)}
\format{
A data frame with 18 observations and 9 variables:
\describe{
  \item{factor1}{experimental factor}
  \item{factor2}{experimental factor}
  \item{factor3}{experimental factor}
  \item{biolrep}{biological replicate}
  \item{techrep}{technical replicates}
  \item{Etarget}{Amplification efficiency of target gene}
  \item{targetCt}{Ct of target gene}
  \item{Eref}{Amplification efficiency of reference gene}
  \item{refCt}{Ct of reference gene}
}
}
\source{
Not applicable
}
\usage{
data_withTechRep
}
\description{
A sample data for calculating biological replicated.
}
\keyword{datasets}
