% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.r
\name{efficiency}
\alias{efficiency}
\title{Slope, R2 and Efficiency (E) statistics and standard curves}
\usage{
efficiency(df)
}
\arguments{
\item{df}{a data frame of dilutions and Ct of genes. First column is dilutions and other columns are Ct values for different genes.}
}
\value{
A list 3 elements.
\describe{
  \item{efficiency}{Slope, R2 and Efficiency (E) statistics}
  \item{Slope_compare}{slope comparison table}
  \item{plot}{standard curves}
  }
}
\description{
The \code{efficiency} function calculates amplification efficiency and returns related statistics and standard curves.
}
\details{
The \code{efficiency} function calculates amplification efficiency of genes, and present the Slope, Efficiency, and R2 statistics.
}
\examples{


# locate and read the sample data
data_efficiency

# Applying the efficiency function
efficiency(data_efficiency)


}
\author{
Ghader Mirzaghaderi
}
