% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.r
\name{efficiency}
\alias{efficiency}
\title{Efficiency, standard curves and related statistics}
\usage{
efficiency(x)
}
\arguments{
\item{x}{a data frame}
}
\value{
A list including standard curves along with the Slope, Efficiency, and R2 statistics
}
\description{
The \code{efficiency} function calculates amplification efficiency and returns standard curves and related statistics.
}
\details{
The \code{efficiency} function calculates amplification efficiency of a target and a reference gene, and present the related standard curves along with the Slope, Efficiency, and R2 statistics. The function also compares the slopes of the two standard curves.
}
\examples{


# locate and read the sample data
data_efficiency

# Applying the efficiency function
efficiency(data_efficiency)


}
\author{
Ghader Mirzaghaderi
}
