% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.r
\docType{data}
\name{Lee_etal2020qPCR}
\alias{Lee_etal2020qPCR}
\title{Sample data (with technical replicates)}
\format{
A data frame with 72 observations and 8 variables:
\describe{
  \item{factor1}{experimental factor}
  \item{DS}{DS}
  \item{biolRep}{biological replicate}
  \item{techRep}{technical replicates}
  \item{APOE_efficiency}{Amplification efficiency of APOE gene}
  \item{APOE_Ct}{Ct of APOE gene}
  \item{GAPDH_efficiency}{Amplification efficiency of GAPDH gene}
  \item{GAPDH_Ct}{Ct of GAPDH gene}
}
}
\source{
Lee et al, (2020) <doi:10.12688/f1000research.23580.2>
}
\usage{
Lee_etal2020qPCR
}
\description{
A sample data for calculating biological replicated.
}
\keyword{datasets}
