% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanTech.r
\name{meanTech}
\alias{meanTech}
\title{Calculating mean of technical replicates}
\usage{
meanTech(x, groups = c(1:4))
}
\arguments{
\item{x}{A raw data frame including technical replicates.}

\item{groups}{grouping columns based on which the mean technical replicates are calculated.}
}
\value{
A data frame with the mean of technical replicates.
}
\description{
Calculating of technical replicates in and output table appropriate for subsequent ANOVA analysis
}
\details{
The meanTech calculates mean of technical replicates.
}
\examples{

# See example input data frame:
data(data_withTechRep)
data_withTechRep

# Calculating mean of technical replicates
meanTech(data_withTechRep, groups = 1:4)


}
\author{
Ghader Mirzaghaderi
}
