% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcrANCOVA.r
\name{qpcrANCOVA}
\alias{qpcrANCOVA}
\title{ANCOVA and ANOVA based on a factorial design plus bar plot of FC}
\usage{
qpcrANCOVA(
  x,
  numberOfrefGenes,
  block = NULL,
  analysisType = "ancova",
  main.factor,
  levels,
  level.names = "none",
  width = 0.5,
  fill = "skyblue",
  y.axis.adjust = 1,
  y.axis.by = 1,
  letter.position.adjust = 0.1,
  ylab = "Average Fold Change",
  xlab = "Pairs",
  fontsize = 12,
  p.adj = c("none", "holm", "hommel", "hochberg", "bonferroni", "BH", "BY", "fdr")
)
}
\arguments{
\item{x}{a data frame. The data frame consists of 4 columns belonging to condition levels, E (efficiency), genes and Ct values, respectively. Each Ct in the following data frame is the mean of technical replicates. Complete amplification efficiencies of 2 is assumed here for all wells but the calculated efficienies can be used we well. We use this data set for Fold Change expression analysis of the target genes in treatment condition compared to normal condition.}

\item{numberOfrefGenes}{number of reference genes. Up to two reference genes can be handled.}

\item{block}{column name of the blocking factor (for correct column arrangement see example data.)}

\item{analysisType}{should be one of "ancova" or "anova".}

\item{main.factor}{main factor (not covariate) for which the levels FC is compared.}

\item{levels}{a numeric vector with the length equal to the factor levels. First number indicates Control.}

\item{level.names}{a vector determining level names in the x axis on the plot.}

\item{width}{a positive number determining bar width.}

\item{fill}{specify the fill color for the columns of the bar plot.}

\item{y.axis.adjust}{a negative or positive value for reducing or increasing the length of the y axis.}

\item{y.axis.by}{determines y axis step length}

\item{letter.position.adjust}{adjust the distance between the signs and the error bars.}

\item{ylab}{the title of the y axis}

\item{xlab}{the title of the x axis}

\item{fontsize}{all fonts size of the plot}

\item{p.adj}{Method for adjusting p values (see p.adjust)}
}
\value{
A list with 2 elements:
\describe{
  \item{Final_data}{}
  \item{lmf}{lm of factorial analysis-tyle}
  \item{lmc}{lm of ANCOVA analysis-type}
  \item{ANOVA_table}{ANOVA table}
  \item{ANCOVA_table}{ANCOVA table}
  \item{Table}{Table of FC values and significance and the 95 percent CI as error bars.}
  \item{plot}{Bar plot of the average fold change for the main factor levels.}
}
}
\description{
ANCOVA (analysis of covariance) and ANOVA (analysis of variance) can be performed using `qpcrANCOVA` function, if more than a factor exist. This works even if there is one factor in the exoeriment. Bar plot of fold changes (FCs) along with the 95\% confidence interval is also returned by the `qpcrANCOVA` function. There is also a function called \code{oneFACTORplot} which returns FC values and related plot for a one-factor-experiment with more than two levels.
}
\details{
The \code{qpcrANCOVA} function applies ANOVA based analysis where one target and one reference gene, that have been evaluated under two or more than two levels of a factor. It returns the bar plot of the average fold change for the target gene along with the 95\% CI and significance.
}
\examples{

# See sample data
data_2factor


qpcrANCOVA(data_2factor, 
           numberOfrefGenes = 1, 
           analysisType = "ancova", 
           main.factor = 2,  
           levels = c(3, 2, 1))
           

qpcrANCOVA(data_2factorBlock, 
           numberOfrefGenes = 1,
           block = "block",  
           analysisType = "ancova", 
           main.factor = 2, 
           levels = c(3, 2, 1))
           
           

}
\author{
Ghader Mirzaghaderi
}
