% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.r
\docType{data}
\name{Taylor_etal2019}
\alias{Taylor_etal2019}
\title{Sample qPCR data (one target and two reference genes under two different conditions)}
\format{
A data frame with 18 observations and 4 variables:
\describe{
  \item{Condition}{Experimental conditions}
  \item{Gene}{Genes}
  \item{E}{Amplification efficiency}
  \item{Ct}{Ct values}
}
}
\source{
Not applicable
}
\usage{
Taylor_etal2019
}
\description{
Sample qPCR data (one target and two reference genes under two different conditions)
}
\keyword{internal}
